requirements.txt
urllib3==1.26.15
python-dateutil==2.8.2
idna==3.4
charset-normalizer==3.1.0
certifi==2022.12.7
six==1.16.0
requests==2.28.2
bs4==0.0.1
beautifulsoup4==4.11.2
soupsieve==2.4
python-dotenv==1.0.0
typing_extensions==4.5.0

psycopg==3.1.8
psycopg-binary==3.1.8

numpy==1.24.2
pytz==2022.7.1
pandas==1.5.3

scikit-learn==1.2.2
scipy==1.10.1
joblib==1.2.0
threadpoolctl==3.1.0

packaging==23.0
Pillow==9.4.0
pyparsing==3.0.9
cycler==0.11.0
kiwisolver==1.4.4
matplotlib==3.7.1